requirement.txt
aiofiles==23.2.1
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
anthropic
anyio
asgiref==3.7.2
Aspose.Slides==24.6.0
asttokens==2.4.1
async-generator==1.10
async-timeout==4.0.3
attrs==22.2.0
Automat==22.10.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.11.1
Brotli==1.1.0
bs4==0.0.1
build==1.1.1
cachetools==5.3.3
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==2.1.1
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
comtypes==1.4.4
constantly==15.1.0
contextlib2==21.6.0
contourpy==1.0.7
cryptography==39.0.1
cssselect==1.2.0
cssselect2==0.7.0
curl_cffi==0.6.2
cycler==0.11.0
dataclasses-json==0.6.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.8.0
docarray==0.32.1
docopt==0.6.2
docx2txt==0.8
duckduckgo_search==5.3.1b1
elementpath==3.0.2
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi
fastjsonschema==2.17.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.12.2
flatbuffers==24.3.7
fonttools==4.39.3
frozenlist==1.4.0
fsspec==2024.2.0
gigachain==0.2.12
gigachain-cli
gigachain-community
gigachain-core
gigachain-text-splitters
gigachat
gigaserve
gitdb==4.0.11
GitPython==3.1.40
google==3.0.0
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.1
grpcio==1.62.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2024.2.26
html5lib==1.1
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx
httpx-sse==0.3.1
huggingface-hub==0.21.4
humanfriendly==10.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
importlib-metadata==6.11.0
importlib_resources==6.3.0
incremental==22.10.0
internetarchive
ipykernel==6.27.1
ipython==8.18.1
isodate==0.6.1
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.3
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.17.3
jupyter_client==8.6.0
jupyter_core==5.5.0
kaleido==0.2.1
kiwisolver==1.4.4
kubernetes
langchain==0.2.14
langchain-anthropic==0.1.19
langchain-cli==0.0.25
langchain-community==0.2.12
langchain-core==0.2.32
langchain-experimental==0.0.63
langchain-openai==0.1.14
langchain-pinecone==0.1.1
langchain-text-splitters==0.2.2
langfuse==2.43.3
langserve==0.2.2
langsmith==0.1.82
libcst==1.4.0
loguru==0.7.2
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
marshmallow==3.20.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
md2pdf==1.0.1
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbformat==5.8.0
nest-asyncio==1.5.8
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.35.10
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
outcome==1.2.0
overrides==7.7.0
packaging==23.2
pandas==1.5.3
parsel==1.8.1
parso==0.8.3
pdfminer.six==20221105
pdfplumber==0.7.6
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.4.0
pinecone-client
platformdirs==4.1.0
plotly==5.14.1
posthog==3.5.0
pptxtopdf==0.0.2
primp==0.5.5
prompt-toolkit==3.0.42
Protego==0.2.1
protobuf==4.25.3
psutil==5.9.6
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.14.5
PyDispatcher==2.0.7
pydyf==0.10.0
Pygments==2.17.2
pyOpenSSL==23.2.0
pyparsing==3.0.9
pypdf==3.17.1
pyphen==0.15.0
PyPika==0.48.9
pyproject-toml==0.0.10
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pyrsistent==0.19.3
pySankey==0.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
python-pptx==0.6.23
pytz==2022.7
PyYAML==6.0.1
pyzmq==25.1.2
queuelib==1.6.2
regex==2024.4.28
requests
requests-file==1.5.1
requests-oauthlib==1.4.0
requests-toolbelt==0.10.1
rich==13.7.0
rsa==4.9
schema==0.7.5
Scrapy==2.9.0
seaborn==0.12.2
selenium
service-identity==23.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.23
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tabula==1.0.5
tabula-py==2.6.0
tavily-python==0.3.3
tenacity==8.2.3
tiktoken==0.7.0
tinycss2==1.3.0
tinysegmenter==0.3
tldextract==3.4.4
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.0
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
typer==0.9.0
types-requests
typing-inspect==0.9.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.19
uvicorn==0.23.2
uvloop==0.19.0
w3lib==2.1.1
Wand==0.6.11
watchfiles==0.21.0
wcwidth==0.2.12
weasyprint==62.3
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
wget==3.2
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
XlsxWriter==3.1.9
xmlschema==2.1.1
yarl==1.9.3
zeep==4.2.1
zipfile36==0.1.3
zipp==3.18.0
zope.interface==6.0
zopfli==0.2.3
